Fixing interval between requests. Increases the waiting time to 60s when error is not OK in the response.

+++ main.py
@@ -36,31 +36,35 @@ while offset < total:
     # do anything with x and perform another request with
     for game in x['results']:
         logging.info("Getting info of the id %s, game: %s" % (game["id"], game["name"]))
-        g = gb.get("/game/" + game["guid"])
-        while g["error"] != 'OK':
-            logging.warning("Sleeping for 5s before trying again.")
-            time.sleep(5)
+        tmp = games.find_one({"_id": game['id'],
+                              'date_last_updated': {"$lte": game['date_last_updated']}})
+        if tmp is None:
+            time.sleep(1)
             g = gb.get("/game/" + game["guid"])
+            while g["error"] != 'OK':
+                logging.warning("Sleeping for 60s before trying again.")
+                time.sleep(60)
+                g = gb.get("/game/" + game["guid"])
 
-        if g["error"] == 'OK':
-            logging.info("Saving game %s" % (game["guid"]))
-            g['results']['_id'] = g['results']['id']
-            g['results']['date_added'] = format_date(g['results']['date_added'])
-            g['results']['date_last_updated'] = format_date(g['results']['date_last_updated'])
-            g['results']['original_release_date'] = format_date(g['results']['original_release_date'])
-            try:
-                games.insert_one(g["results"])
-            except DuplicateKeyError:
-                logging.warning("Trying to insert existing document id %s" % (g['results']['_id']))
-                tmp = games.find_one({"_id": g['results']['id'],
-                                      'date_last_updated': {"$lt": g['results']['date_last_updated']}})
-                if tmp is not None:
-                    logging.warning("Removing old document id %s." % g['results']['id'])
-                    games.delete_one({"_id": g['results']['id']})
-                    logging.warning("Inserting updated document id %s." % g['results']['id'])
+            if g["error"] == 'OK':
+                logging.info("Saving game %s" % (game["guid"]))
+                g['results']['_id'] = g['results']['id']
+                g['results']['date_added'] = format_date(g['results']['date_added'])
+                g['results']['date_last_updated'] = format_date(g['results']['date_last_updated'])
+                g['results']['original_release_date'] = format_date(g['results']['original_release_date'])
+                try:
                     games.insert_one(g["results"])
-                else:
-                    logging.info("There is no update for document id %s." % g['results']['id'])
-                pass
+                except DuplicateKeyError:
+                    logging.warning("Trying to insert existing document id %s" % (g['results']['_id']))
+                    tmp = games.find_one({"_id": g['results']['id'],
+                                          'date_last_updated': {"$lt": g['results']['date_last_updated']}})
+                    if tmp is not None:
+                        logging.warning("Removing old document id %s." % g['results']['id'])
+                        games.delete_one({"_id": g['results']['id']})
+                        logging.warning("Inserting updated document id %s." % g['results']['id'])
+                        games.insert_one(g["results"])
+                    else:
+                        logging.info("There is no update for document id %s." % g['results']['id'])
+                    pass
     offset = offset + limit
     x = gb.get("/games", dict({"offset": offset}))
